removed MySQL as code repository
When loading a new question, instead of querying the MySQL database for the code and pushing it through an "eval(" call, we now load the appropriate script file and append a new script to the body of the document.
This avoids using "eval" an allows browser caching to speed up site. In addition, script files are easier to modify, maintain, fix, etc. The database still holds the question meta-data but this is only queried once when the page loads.